Second attempt to split the code Hopefully this will work
