+++ index.html
@@ -35,7 +35,7 @@
 
   <body>
     <input type="text" placeholder="Message #bot-cmds" id="input" aria-label="Message" autocomplete="off">
-    <script src='main.js'></script>
+    <script src='main.mjs'></script>
   </body>
 
 </html>
