+++ users.mjs
@@ -1,0 +1,3 @@
+import { User } from './constructor.mjs'
+export const browser = new User('You', 'https://yt3.ggpht.com/-7zFDHK5X45w/AAAAAAAAAAI/AAAAAAAAAAA/QJfHeLTEZwE/s900-c-k-no-mo-rj-c0xffffff/photo.jpg')
+export const bot = new User('Bot', './images/bot.png')
